+++ src/components/date_picker/super_date_picker/index.ts
@@ -1,0 +1,15 @@
+export * from './date_popover';
+export * from './quick_select_popover';
+export { AsyncInterval } from './async_interval';
+
+export {
+  EuiSuperDatePicker,
+  EuiSuperDatePickerProps,
+} from './super_date_picker';
+
+export {
+  EuiSuperUpdateButton,
+  EuiSuperUpdateButtonProps,
+} from './super_update_button';
+
+export { prettyDuration, commonDurationRanges } from './pretty_duration';