+++ src/components/index.d.ts
@@ -1,6 +1,3 @@
-/* eslint-disable @typescript-eslint/triple-slash-reference */
-/// <reference path="./date_picker/index.d.ts" />
-
 declare module '@elastic/eui' {
   // @ts-ignore
   export * from '@elastic/eui/src/components/common'; // eslint-disable-line import/no-unresolved